Add files via upload
Building a Recommendation System and a Sentiment Analysis on Amazon Sales Dataset: 
In this project, a recommendation system and a sentiment analysis were built on the Amazon Sales Dataset. First,I built a content-based recommendation system that suggests products based on their similarity to previously purchased products. Then,I performed a sentiment analysis on customer reviews to determine the overall sentiment of the reviews.

This project provides insights into customer preferences and helps businesses make data-driven decisions for product recommendations and marketing strategies.